Add Flash Notices for Failure
When a Task #create or #update operation fails, a flash notice is
presented to the user when the task index page reloads. You can see this
easily by trying to create a new task with the same name as an existing
task.

Additionally, cleaned up controllers to not have unneeded format
blocks.